requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
email-validator==2.1.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
numpy==1.26.4
scikit-learn==1.4.2
openai==1.30.1
pytest==8.1.1
pytest-flask==1.3.0

